+++ src/test/java/com/infomatiq/jsi/DeleteAllEntriesTest.java
@@ -52,8 +52,8 @@ public class DeleteAllEntriesTest extends TestCase {
     Random r = new Random();
     r.setSeed(0);
     for (int i = 0; i < numRects; i += 1) {
-      rects[i] = new Rectangle(r.nextFloat(), r.nextFloat(), r.nextFloat(),
-          r.nextFloat());
+      rects[i] = new Rectangle(r.nextDouble(), r.nextDouble(), r.nextDouble(),
+          r.nextDouble());
     }
 
     run(1, 2, numRects);
@@ -88,13 +88,13 @@ public class DeleteAllEntriesTest extends TestCase {
       rtree.intersects(testRect, counter);
       assertTrue(counter.count == 0);
 
-      rtree.nearest(testPoint, counter, Float.MAX_VALUE);
+      rtree.nearest(testPoint, counter, Double.MAX_VALUE);
       assertTrue(counter.count == 0);
 
-      rtree.nearestN(testPoint, counter, 10, Float.MAX_VALUE);
+      rtree.nearestN(testPoint, counter, 10, Double.MAX_VALUE);
       assertTrue(counter.count == 0);
 
-      rtree.nearestNUnsorted(testPoint, counter, 10, Float.MAX_VALUE);
+      rtree.nearestNUnsorted(testPoint, counter, 10, Double.MAX_VALUE);
       assertTrue(counter.count == 0);
 
       rtree.contains(testRect, counter);
