+++ src/test/java/com/infomatiq/jsi/RTreeWrapper.java
@@ -45,20 +45,20 @@ public class RTreeWrapper implements SpatialIndex {
   }
 
   /**
-   * @see com.infomatiq.jsi.SpatialIndex#nearest(Point, gnu.trove.TIntProcedure, float)
+   * @see com.infomatiq.jsi.SpatialIndex#nearest(Point, gnu.trove.TIntProcedure, double)
    */
   @Override
-  public void nearest(Point p, TIntProcedure v, float furthestDistance) {
+  public void nearest(Point p, TIntProcedure v, double furthestDistance) {
     tree.nearest(new Point(p.x, p.y),
         new IntProcedure2(v),
-        Float.POSITIVE_INFINITY);
+        Double.POSITIVE_INFINITY);
   }
 
   /**
-   * @see com.infomatiq.jsi.SpatialIndex#nearestN(Point, gnu.trove.TIntProcedure, int, float)
+   * @see com.infomatiq.jsi.SpatialIndex#nearestN(Point, gnu.trove.TIntProcedure, int, double)
    */
   @Override
-  public void nearestN(Point p, TIntProcedure v, int n, float furthestDistance) {
+  public void nearestN(Point p, TIntProcedure v, int n, double furthestDistance) {
     tree.nearestN(new Point(p.x, p.y),
         new IntProcedure2(v),
         n,
@@ -66,11 +66,11 @@ public class RTreeWrapper implements SpatialIndex {
   }
 
   /**
-   * @see com.infomatiq.jsi.SpatialIndex#nearestNUnsorted(Point, gnu.trove.TIntProcedure, int, float)
+   * @see com.infomatiq.jsi.SpatialIndex#nearestNUnsorted(Point, gnu.trove.TIntProcedure, int, double)
    */
   @Override
   public void nearestNUnsorted(Point p, TIntProcedure v, int n,
-      float furthestDistance) {
+      double furthestDistance) {
     tree.nearestNUnsorted(new Point(p.x, p.y),
         new IntProcedure2(v),
         n,
